Renamed files, cleaned up some code, started user input
