+++ src/prog_models/models/uav_model/utilities/utils.py
