+++ src/prog_models/models/uav_model/vehicles/control/allocation_functions.py
